% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityMap.R
\docType{package}
\name{activitymap}
\alias{activitymap}
\title{activity: A package using Garmin fit files to plot your activity map.}
\description{
Provides functionality for reading fit files and ploting activity map.
}
