% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertData.R
\name{convertData}
\alias{convertData}
\title{Convert raw data into a list.}
\usage{
convertData(myfiles)
}
\arguments{
\item{myfiles}{raw data read by getData() function}
}
\description{
This function converts raw data into a list, which can be manipulated in R.
adapted from https://msmith.de/FITfileR/articles/FITfileR.html
}
\examples{
convertData(myfiles)
allrecords <- converData(myfiles)
head(allrecords)
allrecords[[1]]
}
