% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Read all .fit files into RStudio.}
\usage{
getData()
}
\description{
This function reads all .fit files in the "data" folder into RStudio.
adapted from https://msmith.de/FITfileR/articles/FITfileR.html
}
\examples{
getData()
myfiles <- getData()
head(myfiles)
myfiles[[1]]
}
