% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionFilter.R
\name{positionFilter}
\alias{positionFilter}
\title{Extract position data and filter wrong position data}
\usage{
positionFilter(allrecords)
}
\arguments{
\item{allrecords}{position records converted from convertData() function}
}
\description{
This function can extract position data and filter wrong position data generated from Garmin devices to modified the final map.
}
\examples{
positionFilter(allrecords)
filtered_position <- positionFilter(allrecords)
head(filtered_position)
}
